EntraceDisplayController: 	Removed the parameters of runDisplays() as I have succesfully maintained the reference of this entranceDisplayController object in the facade all throughout the project.
Changed the return value of findSpotForUser() to boolean instead of void,
in hopes that this will ease our ability to test this method and others that
use it. It does not affect the original logic of the code that the designers
had intended.

ControllerFacade:
Added a bunch of setters and getters to ease the use of the facade when
individual instance variables are needed to be called in AccessControlMain.
Made scanIDCallToController() return a boolean as it may help with verifying
if a user object was succesfully created for the information submitted by the user.

Added startDisplayLogic(), getParkingSpotObject(), currentUserIDlenghtGreaterThan2(),
findSpotForUser(), to hide implementation code of AccessControlMain.java with
facade calls instead.

AccessControlMain:
	Used newly defined facade methods to hide implementation detail
in the main() method.