Update ParkingSpot.java
Created a new method 'sFormat()' in order to be able to test the private method 'format()'. Also made them static in order for me to be able to call them in the ParkingSpotTest.java class. Don't know if that is the correct way to do so, but it got the job done. The new changes DOES NOT change how the program runs... (Again, if you see any errors or concerns with the changes let me know and i will take care of them)